Make name fields nullable and update search logic
Changed several model 'name' fields to nullable in migrations and validation rules for backstamps, effects, patterns, shape collections, colors, and customers. Updated search logic in controllers to remove searching by name fields. Adjusted views to display '-' when name fields are null. Also updated timezone to Asia/Bangkok, changed some seeder counts, and improved dashboard table alignment.

+++ database/migrations/2025_09_25_074234_create_shapes_table.php
@@ -28,13 +28,13 @@ return new class extends Migration
             $table->unsignedBigInteger('designer_id')->nullable();
             $table->unsignedBigInteger('requestor_id')->nullable();
             $table->unsignedBigInteger('updated_by')->nullable();
-            $table->integer('volume')->nullable();
-            $table->integer('weight')->nullable();
-            $table->integer('long_diameter')->nullable();
-            $table->integer('short_diameter')->nullable();
-            $table->integer('height_long')->nullable();
-            $table->integer('height_short')->nullable();
-            $table->integer('body')->nullable();
+            $table->decimal('volume', 8, 2)->nullable();
+            $table->decimal('weight', 8, 2)->nullable();
+            $table->decimal('long_diameter', 8, 2)->nullable();
+            $table->decimal('short_diameter', 8, 2)->nullable();
+            $table->decimal('height_long', 8, 2)->nullable();
+            $table->decimal('height_short', 8, 2)->nullable();
+            $table->decimal('body', 8, 2)->nullable();
             $table->date('approval_date')->nullable();
             $table->timestamps();
